Add support for configuration of section padding for SCTE-104
Provide the ability for a stream option to be set which dictates
whether to do proper iso13818 section padding or to use adaptation
padding (which is supported by some downstream receivers but not
others).

By default we use the legacy behavior (adaptation padding) to
minimize risk, but may wish to revisit this for future releases.