Fix frame dropping if running four encodes in 10-bit mode
The existing code was capturing Y210 and converting to a planer
format, which was then fed to the encoder.  However the encoder
library needed to convert it to NV20 before handing off to the
hardware, which couldn't be sustained at four encodes due to
cache eviction.

Change the colorspace conversion to go directly to NV20, which
avoids the encoder library doing any further colorspace conversion.